Fix warnings on windows

+++ src/ne7ssh.h
@@ -175,7 +175,7 @@ public:
      * @param timeout Timeout for the connection procedure, in seconds.
      * @return Returns newly assigned channel ID, or -1 if connection failed.
      */
-    int connectWithPassword(const char* host, const int port, const char* username, const char* password, bool shell = true, const int timeout = 0);
+    int connectWithPassword(const char* host, const short port, const char* username, const char* password, bool shell = true, const int timeout = 0);
 
     /**
      * Connect to remote host using SSH2 protocol, with publickey authentication.
@@ -189,7 +189,7 @@ public:
      * @param timeout Timeout for the connection procedure, in seconds.
      * @return Returns newly assigned channel ID, or -1 if connection failed.
      */
-    int connectWithKey(const char* host, const int port, const char* username, const char* privKeyFileName, bool shell = true, const int timeout = 0);
+    int connectWithKey(const char* host, const short port, const char* username, const char* privKeyFileName, bool shell = true, const int timeout = 0);
 
     /**
      * Retrieves a pointer to all current connections.
